+++ CHANGELOG.md
@@ -45,6 +45,8 @@ See the [full commit history](https://github.com/DFE-Digital/complete-conversion
 - ProjectGroup is created when a project is edited with a new GRN
 - Added Add, Edit and Delete External Contacts for both conversion and transfer projects.
 - Added `Confirm this transfer has authority to proceed` task
+- Added `Confirm the date the academy transferred` task
+
 - Add service support - view users table
 
 ### Changed